Change wp-login.php to anything you want.
*GoGet Hide Login* is a very light plugin that lets you easily and
safely change the url of the login form to anything you want. It doesn’t
literally rename or change files in core, nor does it add rewrite rules.
It simply intercepts page requests and works on any WordPress website.
The wp-admin directory and wp-login.php page become inaccessible, so you
should bookmark or remember the url. Deactivating this plugin brings
your site back exactly to the state it was before.